+++ tortoise/fields.py
@@ -448,9 +448,12 @@ class BackwardFKRelation(Field):
     __slots__ = ("type", "relation_field")
     has_db_field = False
 
-    def __init__(self, type, relation_field: str) -> None:  # pylint: disable=W0622
+    def __init__(
+        self, type, relation_field: str, description: Optional[str]  # pylint: disable=W0622
+    ) -> None:
         super().__init__(type=type)
         self.relation_field = relation_field
+        self.description = description
 
 
 class RelationQueryContainer: