+++ CHANGELOG.rst
@@ -4,6 +4,14 @@ Changelog
 =========
 0.13.1
 ------
+* Model schema now has a discovery API:
+
+  One can call ``Tortoise.describe_models()`` or ``Tortoise.describe_model(<Model>)`` to get
+  a full description of the model(s).
+
+  Please see :meth:`tortoise.Tortoise.describe_model` and :meth:`tortoise.Tortoise.describe_models` for more info.
+
+- Fix in generating comments for Foreign Keys in ``MySQL``
 - Added schema support for PostgreSQL. Either set  ``"schema": "custom"`` var in ``credentials`` or as a query parameter ``?schema=custom``
 - Default MySQL charset to ``utf8mb4``. If a charset is provided it will also force the TABLE charset to the same.